Send SMS as confirmation on apply
