+++ lib/api/applicantApi.ts
@@ -1,5 +1,10 @@
 import { QueryFunctionContext } from "@tanstack/react-query";
-import { applicantType } from "../types/types";
+import { applicantType, emailDataType } from "../types/types";
+import capitalizeFirstLetter from "../utils/capitalizeFirstLetter";
+import sendEmail from "../email/sendEmail";
+import { generateApplicantEmail } from "../email/applicantEmailTemplate";
+import { changeDisplayName } from "../utils/toString";
+import sendSMS from "../sms/sendSMS";
 
 export const fetchApplicantByPeriodAndId = async (
   context: QueryFunctionContext
