+++ pages/api/applicants/index.ts
@@ -3,13 +3,11 @@ import { createApplicant, getApplicants } from "../../../lib/mongo/applicants";
 import { authOptions } from "../auth/[...nextauth]";
 import { getPeriodById } from "../../../lib/mongo/periods";
 import { getServerSession } from "next-auth";
-import { applicantType, emailDataType } from "../../../lib/types/types";
+import { applicantType } from "../../../lib/types/types";
 import { isApplicantType } from "../../../lib/utils/validators";
 import { isAdmin, hasSession, checkOwId } from "../../../lib/utils/apiChecks";
-import capitalizeFirstLetter from "../../../lib/utils/capitalizeFirstLetter";
-import sendEmail from "../../../lib/email/sendEmail";
-import { changeDisplayName } from "../../../lib/utils/toString";
-import { generateApplicantEmail } from "../../../lib/email/applicantEmailTemplate";
+import { sendConfirmationSMS } from "../../../lib/sms/sendConfirmationSMS";
+import { sendConfirmationEmail } from "../../../lib/email/sendConfirmationEmail";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   const session = await getServerSession(req, res, authOptions);
@@ -55,60 +53,9 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       const { applicant, error } = await createApplicant(requestBody);
       if (error) throw new Error(error);
 
-      if (applicant != null) {
-        let optionalCommitteesString = "";
-        if (applicant.optionalCommittees.length > 0) {
-          optionalCommitteesString = applicant.optionalCommittees
-            ?.map(changeDisplayName)
-            .join(", ");
-        } else {
-          optionalCommitteesString = "Ingen";
-        }
-
-        const emailData: emailDataType = {
-          name: applicant.name,
-          emails: [applicant.email],
-          phone: applicant.phone,
-          grade: applicant.grade,
-          about: applicant.about.replace(/\n/g, "<br>"),
-          firstChoice: "Tom",
-          secondChoice: "Tom",
-          thirdChoice: "Tom",
-          bankom:
-            applicant.bankom == "ja"
-              ? "Ja"
-              : applicant.bankom == "nei"
-              ? "Nei"
-              : "Kanskje",
-          optionalCommittees: optionalCommitteesString,
-        };
-
-        //Type guard
-        if (!Array.isArray(applicant.preferences)) {
-          emailData.firstChoice =
-            applicant.preferences.first == "onlineil"
-              ? "Online IL"
-              : capitalizeFirstLetter(applicant.preferences.first);
-          emailData.secondChoice =
-            applicant.preferences.second == "onlineil"
-              ? "Online IL"
-              : capitalizeFirstLetter(applicant.preferences.second);
-          emailData.thirdChoice =
-            applicant.preferences.third == "onlineil"
-              ? "Online IL"
-              : capitalizeFirstLetter(applicant.preferences.third);
-        }
-
-        try {
-          await sendEmail({
-            toEmails: emailData.emails,
-            subject: "Vi har mottatt din søknad!",
-            htmlContent: generateApplicantEmail(emailData),
-          });
-        } catch (error) {
-          console.error("Error sending email: ", error);
-          throw error;
-        }
+      if (applicant) {
+        await sendConfirmationEmail(applicant);
+        await sendConfirmationSMS(applicant);
       }
 
       return res.status(201).json({ applicant });
